hal/vk: Refactor queue creation by using handles for identifying queue families
